Small update
Unfortunately due to personal issues PR and commit would be smaller but there is still a progress. Now peasants can gather a gold from the mine. Planning to add mine crash and destruction of miners inside if mine has been destroyed + maybe some basic coroutine for peasant to carry wood and gold automatically to the nearest castle of their team.